+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.1] - 2026-2-16
+
+### Improvements
+
+ - 🛠️ Updated track plot window to reset zoom view when axes are changed.
+ - 🛠️ Updated data loader to allow loading a slice of imagery incrementally.
+
 ## [1.10.0] - 2026-2-15
 
 ### Improvements
@@ -285,7 +292,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug where tooltips did not take into account imagery row / column offsets.
 - Fixed bug where imagery produced by algorithms did not have pre-computed histograms (which improves playback performance)
 
-[1.9.2]: https://github.com/awetomaton/VISTA/releases/tag/1.10.0
+[1.10.1]: https://github.com/awetomaton/VISTA/releases/tag/1.10.1
+[1.10.0]: https://github.com/awetomaton/VISTA/releases/tag/1.10.0
 [1.9.1]: https://github.com/awetomaton/VISTA/releases/tag/1.9.1
 [1.9.0]: https://github.com/awetomaton/VISTA/releases/tag/1.9.0
 [1.8.2]: https://github.com/awetomaton/VISTA/releases/tag/1.8.2